mpv: switch to Meson build system
Summary:
Meson build was introduced in mpv 0.35.0, and in this release it's equally supported with waf build. However, waf is going to be dropped in the next release, so we'll have to switch to meson at some point anyway.
Stateless patch was rebased and rewritten to preserve our downstream changes to configuration parsing.

Test Plan:
1. Played a video using mpv.
2. Rebuilt libmpv reverse dependencies: `celluloid`, `haruna`, `mpc-qt` and `vidcutter`.

Reviewers: #triage_team, Staudey

Reviewed By: Staudey

Subscribers: Staudey

Differential Revision: https://dev.getsol.us/D13852